LayoutTests/js/dom/modules/import-incorrect-relative-specifier.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test import rejects the incorrect relative specifiers.');
// Module will be executed asynchronously.
window.jsTestIsAsync = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
<script>
(async function () {
    await shouldReject(`import("incorrect")`);
    await shouldReject(`import("$hello")`);
    await shouldReject(`import(".../test")`);
    finishJSTest();
}());
</script>
</body>
</html>
